Internal Documentation

Documentation for GeneralizedMetropolisHastings.jl's internal functions.

You may need to extend some of these types or functions when developing new problem domains or extending the GMH package.

See Public Documentation for its public interface.

Contents

Internal Documentation

Contents

Index

Internal functions and types

Data - Internal

Index

GeneralizedMetropolisHastings.DataArray

GeneralizedMetropolisHastings.DataFunction

Base.eltype

GeneralizedMetropolisHastings.datatypename

GeneralizedMetropolisHastings.generate!

Internal functions and types

Data - Internal

Data - Internal Functions

GeneralizedMetropolisHastings.datatypename — Method.

datatypename(d)

Return the name of the data type (used by show()).

source

Base.eltype — Method.

eltype(d)

Return the type of the data values.

source

GeneralizedMetropolisHastings.generate! — Method.

generate!(d)

Generate a set of data values. The function should return its argument d.

source

Data - Internal Types

GeneralizedMetropolisHastings.DataArray — Type.

DataArray

A type wrapping a predefined array of measurement data.

source

GeneralizedMetropolisHastings.DataFunction — Type.

DataFunction

A type wrapping a data generating function.

source