Public Documentation

Documentation for GeneralizedMetropolisHastings.jl's public interface.

See Internal Documentation for internal package docs.

Contents

Public Documentation

Contents

Index

Public Interface

Data - Interface

Index

GeneralizedMetropolisHastings.AbstractData

GeneralizedMetropolisHastings.data

GeneralizedMetropolisHastings.dataindex

GeneralizedMetropolisHastings.datavalues

GeneralizedMetropolisHastings.numvalues

GeneralizedMetropolisHastings.numvars

Public Interface

Data - Interface

Data - Public Functions

GeneralizedMetropolisHastings.data — Function.

data(s::Symbol,args...)

Create data objects.

Currently implemented variants:

source

data(:array,index::AbstractVector,values::AbstractArray)

Create a data object wrapping an array of fixed, predefined data values.

Arguments

index::AbstractVector: an index for each row of the values array (e.g., a time index)

values::AbstractArray: each row is a data point and each column a variable

source

data(:function!,index::AbstractVector,out::AbstractArray,f!::Function,[,...args])

Create a DataFunction object which encapsulates an in-place data generating function.

The :function! variant expects a pre-allocated output array out, and a data generating function which takes out as its first argument.

Arguments

index::AbstractVector: an index for the rows of the result (e.g., a time index)

out::AbstractArray: array with one row per data point and one column per variable

f!::Function: a data generating function taking as first argument out

args...: additional arguments of the data generating function

source

data(:function,index::AbstractVector,f::Function[,args...])

Create a DataFunction object which encapsulates a data generating function.

The :function variant can be used if the data generating function cannot generate its result in a pre-allocated output array. The result should contain one row per data point and one column per variable.

Arguments

index::AbstractVector: an index for the rows of the result (e.g., a time index)

f::Function: a function returning an array of values as a result

args...: additional arguments of the data generating function

source

GeneralizedMetropolisHastings.dataindex — Method.

dataindex(d)

Retrieve the data index.

source

GeneralizedMetropolisHastings.datavalues — Method.

datavalues(d)

Retrieve the data values.

source

GeneralizedMetropolisHastings.numvalues — Method.

numvalues(d)

Return the number of data values.

source

GeneralizedMetropolisHastings.numvars — Method.

numvars(d)

Return the number of variables.

source

Data - Public Types

GeneralizedMetropolisHastings.AbstractData — Type.

Base type for objects wrapping measurement data.

source